Make it possible to influence format preference order.
